+++ src/Continuation/State.php
@@ -1,0 +1,13 @@
+<?php
+declare(strict_types = 1);
+
+namespace Innmind\FileWatch\Continuation;
+
+/**
+ * @psalm-immutable
+ */
+enum State
+{
+    case continue;
+    case stop;
+}